[Chore #163356402] Refactor code
